requirements.txt
pillow==7.2.0
tensorflow==2.3.0
matplotlib==3.3.1
seldon-core==1.14.0